Added the ProtocolCommentary and ProtocolNote entities and associated them with their respective entities. Updated the API endpoints which ad/update comments and notes on protocols and protocol steps. This update maintains the existing Update operation which "overwrites" notes and commentary. There are "append" endpoints for each which should be used if one looks to append commentary or notes. Update operations overwrite data. The notes and commentary additions were then propagated to the respective endpoints to update JSON which represents the notes and commentary now as an array of Objects as opposed to simple string field.
Added sub-category functionality to step templates. Updated the GetAllStepTemplateCategoriesEndpoint to represent the parent/child relationships in the Categories. Added two new endpoints which allow users to GetAllStepTemplateCategoriesByParentCategoryId and GetAllStepTemplatesByTemplateCategoryId.

Of course the updated Liquibase changes are added also which migrate the existing comments and notes to the respective tables and delete the comments and notes columns from the protocol and protocolStep tables respectively.